+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ## Bug Fixes
  * Some JMS Consumers and Producers are filtered due to class name filtering in instrumentation matching
  * Jetty: When no display name is set and context path is "/" transaction service names will now correctly fall back to configured values
+ * Drops non-String labels when connected to APM Server < 6.7 to avoid validation errors (#687)
 
 # 1.7.0